# apps/backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: Gestiabloc API
  version: 1.0.0
  description: |
    Authentication and file upload endpoints for Gestiabloc.
servers:
  - url: http://localhost:4000
    description: Local backend
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user (password-based)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkWithUserId"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/tenant/signup:
    post:
      tags: [Auth]
      summary: Public tenant signup
      description: Creates tenant, subscription, admin user, and membership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantSignupRequest"
      responses:
        "201":
          description: Tenant + admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSignupResponse"
        "400":
          description: Invalid tenant name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (step 1) and issue OTP challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP (step 2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Invalid challenge or OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
    patch:
      tags: [Auth]
      summary: Update current user profile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me/password:
    patch:
      tags: [Auth]
      summary: Update current user password
      description: Requires current password and enforces 8+ chars, at least one letter, one number, and one symbol.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid current password or invalid new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Start password reset
      description: Always returns ok to avoid account enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
...
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
    OkWithUserId:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        userId:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "INVALID_CREDENTIALS"
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    TenantSignupRequest:
      type: object
      required: [tenantName, planKey, adminName, adminEmail, adminPassword]
      properties:
        tenantName:
          type: string
        planKey:
          type: string
          enum: [STARTER, PRO, BUSINESS]
        paidNow:
          type: boolean
          default: false
        adminName:
          type: string
        adminEmail:
          type: string
          format: email
        adminPassword:
          type: string
          minLength: 8
    TenantSignupResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        tenantId:
          type: string
        userId:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        requiresOtp:
          type: boolean
        challengeToken:
          type: string
          nullable: true
    OtpVerifyRequest:
      type: object
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
        code:
          type: string
          description: 6-digit OTP
          pattern: "^[0-9]{6}$"
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
          description: Must include at least one letter, one number, and one symbol.
    UpdateMeRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 255
    UpdatePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 200
        newPassword:
          type: string
          minLength: 8
          maxLength: 200
          description: Must include at least one letter, one number, and one symbol.
    MeResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
              format: email
            name:
              type: string
            image:
              type: string
              nullable: true
            platformRole:
              type: string
            emailVerified:
              type: boolean
            lastLoginAt:
              type: string
              format: date-time
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            memberships:
              type: array
              items:
                type: object
                properties:
                  role:
                    type: string
                  status:
                    type: string
                  securityLevel:
                    type: string
                    enum: [LOW, MEDIUM, MAX]
                  tenant:
                    type: object
                    properties:
                      id:
                        type: string
                      slug:
                        type: string
                      name:
                        type: string
...